+++ client/src/components/views/PostListPage/PostDetail.jsx
@@ -1,121 +1,234 @@
-import React, { useEffect, useState } from "react";
-import styled from "styled-components";
+import React, { useState, useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
+import styled from "styled-components";
+import backarrowwhite from "../../images/backarrow-white.svg";
+import moredetail from "../../images/moredetails.svg";
 import axios from "axios";
-import backarrow from "../../images/backarrow.svg";
+import api from "../../utils/api";
 
-// PostDetail 컴포넌트
+import { Swiper, SwiperSlide } from "swiper/react";
+import "swiper/css";
+import "swiper/css/pagination";
+import { Pagination } from "swiper";
+import { createGlobalStyle } from "styled-components";
+
+const GlobalStyle = createGlobalStyle`
+  .swiper-pagination-bullet {
+    background-color: black !important; /* 기본 점박이 빨간색 */
+    opacity: 1;
+  }
+
+  .swiper-pagination-bullet-active {
+    background-color: red !important; /* 활성화된 점박이 검정색 */
+    opacity: 1;
+  }
+`;
+
 function PostDetail() {
-  const { sellId } = useParams(); // URL에서 sellId 가져오기
-  console.log("현재 sellId:", sellId);
-
+  const { sellId } = useParams();
   const navigate = useNavigate();
-  const [sell, setSell] = useState(null);
-  const [exchangeRate, setExchangeRate] = useState(null);
+  const [sell, setSell] = useState({});
+  const [exchangeRates, setExchangeRates] = useState({});
+  const [latitude, setLatitude] = useState(null);
+  const [longitude, setLongitude] = useState(null);
+  const [showMenu, setShowMenu] = useState(false);
+
+  const toggleMenu = () => {
+    setShowMenu((prevState) => !prevState);
+  };
+
+  const handleEdit = () => {
+    alert("수정");
+    setShowMenu(false);
+  };
+
+  const handleDelete = () => {
+    alert("삭제");
+    setShowMenu(false);
+  };
 
   useEffect(() => {
-    console.log("현재 sellId:", sellId);  //  sellId 값 확인
     if (!sellId) {
-        console.error("sellId가 undefined입니다.");
-        return;
+      console.error("sellId가 undefined입니다.");
+      return;
     }
-
+  
     const fetchPost = async () => {
       try {
-        // ✅ 토큰 가져오기 (localStorage 또는 sessionStorage)
-        const accessToken = localStorage.getItem("accessToken") || sessionStorage.getItem("accessToken");
-        console.log("현재 저장된 accessToken:", accessToken);
-
-        if (!accessToken) {
-          alert("로그인이 필요합니다.");
-          navigate("/login"); // 로그인 페이지로 이동
-          return;
-        }
-
-        const response = await axios.get(`http://localhost:5000/sell/sellDescription/${sellId}`, {
-          headers: {
-            "Content-Type": "application/json", // ✅ 수정된 부분
-            Authorization: `Bearer ${accessToken}`, 
-          },
-          withCredentials: true,
-        });
-
-        console.log("불러온 판매 데이터:", response.data);
-        setSell(response.data);
+        const response = await api.get(`/api/sell/sellDescription/${sellId}`);
+        console.log("서버에서 받은 데이터:", response.data);
+        setSell(response.data || {});
+        
+        // 판매 정보에서 좌표를 바로 설정
+        setLatitude(response.data.latitude);
+        setLongitude(response.data.longitude);
       } catch (error) {
         console.error("판매 정보 불러오기 실패:", error);
         if (error.response?.status === 401) {
           alert("인증이 만료되었습니다. 다시 로그인해주세요.");
-          navigate("/login"); 
+          navigate("/login");
         } else if (error.response?.status === 404) {
           alert("판매 정보를 찾을 수 없습니다.");
         }
       }
     };
-
+  
     fetchPost();
   }, [sellId, navigate]);
 
-  if (!sell) {
+  //  실시간 환율 가져오기
+  useEffect(() => {
+    const fetchExchangeRates = async () => {
+      try {
+        const response = await axios.get(`https://api.exchangerate-api.com/v4/latest/${sell.currency}`);
+        setExchangeRates((prevRates) => ({
+          ...prevRates,
+          [sell.currency]: response.data.rates.KRW,
+        }));
+      } catch (error) {
+        console.error("환율 데이터 불러오기 오류:", error);
+      }
+    };
+  
+    if (sell.currency) {
+      fetchExchangeRates();
+    }
+  }, [sell]);
+  
+  
+
+  //판매자 거래 희망 장소 카맵에 마커로 띄우기
+  useEffect(() => {
+    if (latitude && longitude) {  // 좌표가 있을 때만 지도 로드
+      const script = document.createElement("script");
+      script.src = `https://dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.REACT_APP_KAKAOMAP_KEY}&autoload=false`;
+      script.async = true;
+  
+      script.onload = () => {
+        window.kakao.maps.load(() => {
+          const container = document.getElementById("kakao-map");
+          const options = {
+            center: new window.kakao.maps.LatLng(latitude, longitude),
+            level: 3,
+          };
+  
+          const map = new window.kakao.maps.Map(container, options);
+          new window.kakao.maps.Marker({
+            position: new window.kakao.maps.LatLng(latitude, longitude),
+            map: map,
+          });
+        });
+      };
+  
+      document.body.appendChild(script);
+    }
+  }, [latitude, longitude]);
+  
+
+  if (!sell || Object.keys(sell).length === 0) {
     return <LoadingMessage>데이터를 불러오는 중...</LoadingMessage>;
   }
 
   return (
-    <Container>
-      
-      <ImageBackground>
-        {sell.images && sell.images.length > 0 ? (
-          <MainImage src={sell.images[0]} alt="상품 이미지" />
-        ) : (
-          <NoImage>이미지 없음</NoImage>
-        )}
-        <TopBar>
-          <BackButton onClick={() => navigate(-1)} src={backarrow} alt="뒤로가기" />
-        </TopBar>
-      </ImageBackground>
+    <>
+      <GlobalStyle />
+      <Container>
+        <ImageBackground>
+          {sell?.images && sell.images.length > 0 ? (
+            <Swiper
+              modules={[Pagination]}
+              pagination={{
+                clickable: true,
+              }}
+              spaceBetween={10}
+              slidesPerView={1}
+              loop={true}
+            >
+              {sell.images.map((image, index) => (
+                <SwiperSlide key={index}>
+                  <MainImage
+                    src={image}
+                    alt={`상품 이미지 ${index + 1}`}
+                    onError={(e) => (e.target.src = "/fallback-image.png")}
+                  />
+                </SwiperSlide>
+              ))}
+            </Swiper>
+          ) : (
+            <NoImage>이미지 없음</NoImage>
+          )}
 
-   
-      <Content>
-        <CurrencyTag>{sell.currency}</CurrencyTag>
-        <Price>${sell.amount.toLocaleString()}</Price>
+          <TopBar>
+            <BackButton
+              onClick={() => window.history.back()}
+              src={backarrowwhite}
+              alt="뒤로가기"
+            />
+            <MenuButton onClick={toggleMenu} src={moredetail} alt="더보기" />
+          </TopBar>
+          {showMenu && (
+            <Menu>
+              <MenuItem onClick={handleEdit}>수정</MenuItem>
+              <MenuItem onClick={handleDelete}>삭제</MenuItem>
+            </Menu>
+          )}
+        </ImageBackground>
 
-        <InfoSection>
-          <InfoTitle>거래 위치</InfoTitle>
-          <InfoValue>{sell.location || "위치 정보 없음"}</InfoValue>
-        </InfoSection>
+        <Content>
+          <TopInfo>
+            <CurrencyTag>{sell.currency}</CurrencyTag>
+            <UserInfo>
+            <UserImage
+              src={sell.profile_img || "https://via.placeholder.com/40"} 
+              alt="판매자 프로필"
+            />
+              <UserName>{sell.name || "익명 판매자"}</UserName>
+            </UserInfo>
+          </TopInfo>
+          <Price>${sell.amount?.toLocaleString()}</Price>
+          <InfoSection>
+            <InfoTitle>거래 위치</InfoTitle>
+            <InfoValue>{sell.location || "위치 정보 없음"}</InfoValue>
+          </InfoSection>
+          <InfoSection>
+            <InfoTitle>환율</InfoTitle>
+            <InfoValue>
+              {exchangeRates[sell.currency]
+              ? `100 ${sell.currency} / ${exchangeRates[sell.currency].toFixed(2)} 원`
+              : "환율 정보 없음"}
+            </InfoValue>
 
-        <InfoSection>
-          <InfoTitle>환율</InfoTitle>
-          <InfoValue>
-            {exchangeRate ? `100 ${sell.currency} / ${exchangeRate.toFixed(2)} 원` : "환율 정보 없음"}
-          </InfoValue>
-        </InfoSection>
+          </InfoSection>
+          <Description>{sell.content || "설명 없음"}</Description>
+          <LocationInfo>
+            <LocationTitle>거래 희망 장소</LocationTitle>
+            <LocationAddress>{sell.location}</LocationAddress>
+          </LocationInfo>
 
-        <Description>{sell.content || "설명 없음"}</Description>
+          <MapContainer
+            id="kakao-map"
+            style={{ width: "100%", height: "250px" }}
+          ></MapContainer>
 
-        <UserInfo>
-          <UserImage src="https://via.placeholder.com/40" alt="사용자 프로필" />
-          <UserName>{sell.name || "익명 판매자"}</UserName>
-        </UserInfo>
-
-    
-        <KRWContainer>
-          <KRWLabel>원화</KRWLabel>
-          <KRWAmount>
-            {exchangeRate ? `${Math.round(sell.amount * exchangeRate).toLocaleString()} 원` : "환율 정보 없음"}
-          </KRWAmount>
-        </KRWContainer>
-
-      
-        <InquiryButton>문의하기</InquiryButton>
-      </Content>
-    </Container>
+          <ButtonContainer>
+            <KRWContainer>
+              <KRWLabel>원화</KRWLabel>
+              <KRWAmount>
+                {exchangeRates[sell.currency]
+                ? `${Math.round(sell.amount * exchangeRates[sell.currency]).toLocaleString()} 원`
+                : "환율 정보 없음"}
+                </KRWAmount>
+            </KRWContainer>
+            <InquiryButton>문의하기</InquiryButton>
+          </ButtonContainer>
+        </Content>
+      </Container>
+    </>
   );
 }
 
 export default PostDetail;
 
-// 스타일 정의
 const Container = styled.div`
   width: 100%;
   height: 100vh;
@@ -126,13 +239,13 @@ const Container = styled.div`
 const ImageBackground = styled.div`
   position: relative;
   width: 100%;
-  height: 250px;
+  height: 350px;
   background-color: #f0f0f0;
 `;
 
 const MainImage = styled.img`
   width: 100%;
-  height: 100%;
+  height: 350px;
   object-fit: cover;
 `;
 
@@ -145,41 +258,75 @@ const NoImage = styled.div`
   color: #999;
 `;
 
+
 const TopBar = styled.div`
   position: absolute;
   top: 10px;
-  left: 10px;
-  right: 10px;
+  left: 0;
+  right: 0;
   display: flex;
   justify-content: space-between;
+  z-index: 10;
+  width: 100%; /* 전체 너비 차지 */
 `;
 
 const BackButton = styled.img`
-  width: 24px;
-  height: 24px;
+  width: 20px;
+  height: 20px;
   cursor: pointer;
+  margin-left: 0;
 `;
 
-const ShareButton = styled.img`
-  width: 24px;
-  height: 24px;
+const MenuButton = styled.img`
+  width: 20px;
+  height: 20px;
   cursor: pointer;
+  margin-right: 0;
 `;
 
 const Content = styled.div`
   padding: 20px;
   background: white;
   flex: 1;
+  margin: 0 auto; /* 가운데 정렬 */
+  width: 100%; /* 전체 너비 차지 */
+`;
+
+const TopInfo = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 10px;
+  width: 100%;
 `;
 
 const CurrencyTag = styled.div`
-  background: red;
+  background: #ca2f28;
   color: white;
   font-size: 14px;
-  font-weight: bold;
+  font-weight: 300;
   padding: 5px 10px;
   border-radius: 5px;
   display: inline-block;
+  margin-left: 0;
+`;
+
+const UserInfo = styled.div`
+  display: flex;
+  align-items: center;
+  margin-right: 0;
+`;
+
+const UserImage = styled.img`
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  margin-right: 10px;
+`;
+
+const UserName = styled.span`
+  font-size: 16px;
+  font-weight: 500;
 `;
 
 const Price = styled.h1`
@@ -193,16 +340,48 @@ const InfoSection = styled.div`
   justify-content: space-between;
   border-bottom: 1px solid #eee;
   padding: 10px 0;
+  margin-top: 13px;
 `;
 
 const InfoTitle = styled.span`
   font-size: 14px;
-  color: #555;
+  color: #1F2024;
+  margin-left: 0;
 `;
 
 const InfoValue = styled.span`
   font-size: 14px;
+  font-weight: 200;
+  color: #666;
+  margin-right: 0;
+`;
+
+const LocationInfo = styled.div`
+  margin-top: 20px;
+  padding: 10px;
+  border-radius: 10px;
+  margin-left: -10px;
+  width: 100%;
+`;
+
+const LocationTitle = styled.h3`
+  font-size: 17px;
   font-weight: bold;
+  color: #333;
+  margin-left: 0;
+`;
+
+const LocationAddress = styled.p`
+  font-size: 13px;
+  color: #898d99;
+  margin-top: 5px;
+  margin-left: 0;
+`;
+
+const LoadingMessage = styled.div`
+  text-align: center;
+  margin-top: 20px;
+  color: #666;
 `;
 
 const Description = styled.p`
@@ -211,55 +390,84 @@ const Description = styled.p`
   color: #666;
 `;
 
-const UserInfo = styled.div`
+const Menu = styled.div`
+  position: absolute;
+  top: 40px;
+  right: 10px;
+  background-color: white;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+  z-index: 10;
+`;
+
+const MenuItem = styled.div`
+  padding: 10px;
+  font-size: 14px;
+  cursor: pointer;
+  &:hover {
+    background-color: #f5f5f5;
+  }
+`;
+
+const ButtonContainer = styled.div`
+  position: fixed;
+  bottom: 0;
+  left: 50%;
+  transform: translateX(-50%);
   display: flex;
+  justify-content: center;
   align-items: center;
-  margin-top: 20px;
-`;
-
-const UserImage = styled.img`
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  margin-right: 10px;
-`;
-
-const UserName = styled.span`
-  font-size: 16px;
-  font-weight: bold;
+  width: 90%;
+  background: white;
+  padding: 10px;
+  z-index: 10;
 `;
 
 const KRWContainer = styled.div`
   margin-top: 20px;
   padding: 15px;
-  background: #f5f5f5;
   border-radius: 10px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: column;
+  border-top: 1px solid #eee;
+  justify-content: flex-start;
+  margin-left: 0;
 `;
 
 const KRWLabel = styled.span`
   font-size: 14px;
   color: #888;
+  margin-left: -10px;
+  justify-content: flex-start;
 `;
 
 const KRWAmount = styled.h2`
-  font-size: 22px;
-  font-weight: bold;
+  font-size: 13px;
+  font-weight: 300;
+  margin-left: -10px;
 `;
 
 const InquiryButton = styled.button`
-  width: 100%;
+  width: 250px;
   padding: 15px;
-  margin-top: 20px;
-  background: black;
+  background: #1f2024;
   color: white;
   font-size: 16px;
-  font-weight: bold;
+  font-weight: 500;
   border: none;
   border-radius: 10px;
   cursor: pointer;
+  align-self: flex-end;
+  border-top: 1px solid #eee;
+  margin-right:0;
 `;
-const LoadingMessage = styled.div`
-  text-align: center;
+
+const MapContainer = styled.div`
   margin-top: 20px;
-  color: #666;
+  width: 100%;
+  height: 300px;
+  margin-left: 0;
 `;
